transaction api fixed

+++ src/transaction/transaction.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreateTransactionDto } from './dto/create-transaction.dto';
 import { UpdateTransactionDto } from './dto/update-transaction.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
@@ -20,10 +20,17 @@ export class TransactionService {
     return 'Transaction created successfully';
   }
 
-  async findAll(page: number = 1, total: number = 10) {
+  async findAll(page: number = 1, limit: number = 10) {
+    const paginate: { skip: number; take: number } | {} =
+      Number(page) && Number(limit)
+        ? {
+            skip: (Number(page) - 1) * Number(limit),
+            take: Number(limit),
+          }
+        : {};
+
     const transactions = await this.prisma.transaction.findMany({
-      skip: (page - 1) * total,
-      take: total,
+      ...paginate,
       orderBy: {
         date: 'desc',
       },
@@ -36,18 +43,47 @@ export class TransactionService {
     const transaction = await this.prisma.transaction.findUnique({
       where: { id },
     });
+    if (!transaction) {
+      throw new HttpException(
+        `Bad Request. No transaction found with id: ${id}`,
+        HttpStatus.BAD_REQUEST,
+      );
+    }
     return transaction;
   }
 
   async update(id: number, updateTransactionDto: UpdateTransactionDto) {
-    await this.prisma.transaction.update({
+    const findTransaction = await this.findOne(id);
+    if (!findTransaction) {
+      throw new HttpException(
+        `Bad Request. No transaction found with id: ${id}`,
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
+    const transaction = await this.prisma.transaction.update({
       where: { id },
       data: updateTransactionDto,
     });
+
+    if (!transaction) {
+      throw new HttpException(
+        `Error!. Transaction could not be updated with id: ${id}`,
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     return 'Transaction updated successfully';
   }
 
   async remove(id: number) {
+    const findTransaction = await this.findOne(id);
+    if (!findTransaction) {
+      throw new HttpException(
+        `Bad Request. No transaction found with id: ${id}`,
+        HttpStatus.BAD_REQUEST,
+      );
+    }
     await this.prisma.transaction.delete({ where: { id } });
     return 'Transaction deleted successfully';
   }
